modularized and finished code (endpoint complete)
