+++ .envrc.example
@@ -1,0 +1,2 @@
+export SPOONACULAR_URL="XXXXXXXXXX"
+export SPOONACULAR_API_KEY="XXXXXXXXXXXXXXX"
