+++ .gitignore
@@ -120,7 +120,7 @@ celerybeat.pid
 *.sage.py
 
 # Environments
-.env
+.envrc
 .venv
 env/
 venv/
